requirements.txt
# PyTorch (CUDA 11.8)
torch==2.2.2+cu118
torchvision==0.17.2+cu118
torchaudio==2.2.2+cu118
--extra-index-url https://download.pytorch.org/whl/cu118

# 你指定的版本（去掉 ^）
diffusers==0.31.0
transformers==4.55.3
tokenizers==0.21.0
huggingface_hub==0.34.4
accelerate==1.10.0
safetensors==0.4.5
requests==2.32.3
pyyaml==6.0.2

# GUI / 影像
Pillow==10.4.0
